update:djangochannelsrestframework使用 POSTMANでテスト済
