+++ playground/rtakashi/djangochannelsrestframework/websocket/models.py
@@ -11,5 +11,5 @@ class User(AbstractUser):
 class GameStatus(models.Model):
     game_status = models.CharField(max_length=10)
 
-    def __str__(self)
+    def __str__(self):
         return f"GameStatus({self.id} {self.game_status})"
