+++ playground/rtakashi/djangochannelsrestframework/websocket/routing.py
@@ -1,7 +1,7 @@
 from django.urls import path
-
 from .PongLogic import consumers
 
 websocket_urlpatterns = [
     path("ws/websocket/", consumers.PongLogic.as_asgi()),
+    path("ws/api/gamestatus/", consumers.GameStatusConsumer.as_asgi()),
 ]
