+++ playground/rtakashi/djangochannelsrestframework/websocket/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+
+
+class GameStatusSerializer(serializers.ModelSerializer):
+    class Meta:
+        from .models import GameStatus
+        model = GameStatus
+        fields = ['id', 'game_status']
