+++ playground/rtakashi/djangochannelsrestframework/websocket/templates/websocket/index.html
