+++ playground/rtakashi/djangochannelsrestframework/websocket/urls.py
