--- playground/rtakashi/djangochannelsrestframework/websocket_rtakashi/requirements.txt
@@ -1,8 +1,0 @@
-asgiref
-django
-django-otp
-sqlparse
-channels
-channels_redis
-daphne
-djangorestframework
