--- playground/rtakashi/djangochannelsrestframework/websocket_rtakashi/websocket/serializers.py
@@ -1,25 +1,0 @@
-from django.contrib.auth.models import Group, User
-from rest_framework import serializers
-
-
-class UserSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = User
-        fields = ['url', 'username', 'email', 'groups']
-
-
-class GroupSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = Group
-        fields = ['url', 'name']
-
-# 以下を追加
-from rest_framework import serializers
-from .models import GameStatus
-
-
-class GameStatusSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = GameStatus
-        fields = ['id', 'status']
-
