--- playground/rtakashi/djangochannelsrestframework/websocket_rtakashi/websocket_rtakashi/urls.py
@@ -1,43 +1,0 @@
-"""
-URL configuration for websocket_rtakashi project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/5.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-
-from django.contrib import admin
-# from django.urls import include, path
-
-# urlpatterns = [
-#     path("admin/", admin.site.urls),
-#     path("", include("websocket.urls")),
-#     path('api-auth/', include('rest_framework.urls'))
-# ]
-
-
-from django.urls import include, path
-from rest_framework import routers
-
-from websocket import views
-
-# router = routers.DefaultRouter()
-# router.register(r'users', views.UserViewSet)
-# router.register(r'groups', views.GroupViewSet)
-
-# Wire up our API using automatic URL routing.
-# Additionally, we include login URLs for the browsable API.
-urlpatterns = [
-    path("admin/", admin.site.urls),
-    path("", include("websocket.urls")),
-    path('gamestatus/', include('websocket.urls'))
-]
