--- playground/rtakashi/djangochannelsrestframework/websocket_rtakashi/websocket_rtakashi/wsgi.py
@@ -1,16 +1,0 @@
-"""
-WSGI config for websocket_rtakashi project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/5.1/howto/deployment/wsgi/
-"""
-
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "websocket_rtakashi.settings")
-
-application = get_wsgi_application()
